+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-## 0.5.0
+## 0.5.1
 
-- adding support for different Gemini models (gemini-1.5-flash, gemini-2.0-flash, gemini-2.0-flash-lite, gemini-2.5-pro, gemini-2.5-flash)
+- updating the models to support the gemini 3 models.
+- debugging hints added.